refactor(data-warehouse): split sql editor scene logic
