+++ frontend/src/scenes/data-warehouse/editor/EditorScene.tsx
@@ -1,25 +1,33 @@
 import './EditorScene.scss'
 
+import { useActions } from 'kea'
+
 import { AccessDenied } from 'lib/components/AccessDenied'
 import { userHasAccess } from 'lib/utils/accessControlUtils'
 import { SceneExport } from 'scenes/sceneTypes'
 
 import { AccessControlLevel, AccessControlResourceType } from '~/types'
 
+import { editorSceneLogic } from './editorSceneLogic'
 import { SQLEditor } from './SQLEditor'
-import { sqlEditorLogic } from './sqlEditorLogic'
 import { SQLEditorMode } from './sqlEditorModes'
 
 export const scene: SceneExport = {
-    logic: sqlEditorLogic,
+    logic: editorSceneLogic,
     component: EditorScene,
 }
 
 export function EditorScene({ tabId }: { tabId?: string }): JSX.Element {
+    const resolvedTabId = tabId ?? 'default'
+    const { shareTab } = useActions(editorSceneLogic({ tabId: resolvedTabId }))
+
     if (!userHasAccess(AccessControlResourceType.WarehouseObjects, AccessControlLevel.Viewer)) {
         return (
             <AccessDenied reason="You don't have access to Data warehouse tables & views, so the SQL editor isn't available." />
         )
     }
-    return <SQLEditor tabId={tabId} mode={SQLEditorMode.FullScene} showDatabaseTree={true} />
+
+    return (
+        <SQLEditor tabId={resolvedTabId} mode={SQLEditorMode.FullScene} showDatabaseTree={true} onShareTab={shareTab} />
+    )
 }
