+++ frontend/src/scenes/data-warehouse/editor/OutputPane.tsx
@@ -399,7 +399,7 @@ interface ResultsActionsProps {
     exportContext: ExportContext | undefined
     hasQueryInput: boolean
     isEmbeddedMode: boolean
-    onShareTab: () => void
+    onShareTab?: () => void
 }
 
 function ResultsActions({
@@ -457,7 +457,7 @@ function ResultsActions({
                     />
                 </Tooltip>
             )}
-            {!isEmbeddedMode && (
+            {!isEmbeddedMode && onShareTab && (
                 <Tooltip title="Share your current query">
                     <LemonButton
                         id="sql-editor-share"
@@ -482,7 +482,7 @@ interface OutputActionsProps {
     hasQueryInput: boolean
     isEmbeddedMode: boolean
     settingsOpen: boolean
-    onShareTab: () => void
+    onShareTab?: () => void
     onToggleChartSettingsPanel: () => void
 }
 
@@ -528,14 +528,15 @@ function OutputActions({
 interface OutputPaneProps {
     tabId: string
     showToolbar?: boolean
+    onShareTab?: () => void
 }
 
-export function OutputPane({ tabId, showToolbar = true }: OutputPaneProps): JSX.Element {
+export function OutputPane({ tabId, showToolbar = true, onShareTab }: OutputPaneProps): JSX.Element {
     const { activeTab } = useValues(outputPaneLogic)
     const { setActiveTab } = useActions(outputPaneLogic)
 
     const { sourceQuery, exportContext, insightLoading, hasQueryInput, isEmbeddedMode } = useValues(sqlEditorLogic)
-    const { setSourceQuery, shareTab } = useActions(sqlEditorLogic)
+    const { setSourceQuery } = useActions(sqlEditorLogic)
     const { isDarkModeOn } = useValues(themeLogic)
     const {
         response: dataNodeResponse,
@@ -750,7 +751,7 @@ export function OutputPane({ tabId, showToolbar = true }: OutputPaneProps): JSX.
         hasQueryInput,
         isEmbeddedMode,
         settingsOpen: isChartSettingsPanelOpen,
-        onShareTab: shareTab,
+        onShareTab,
         onToggleChartSettingsPanel: toggleVisualizationSettingsPanel,
     }
